requirements.txt
# Core compositor dependencies
numpy
pillow
pycairo
pytoshop
six

# REST API
fastapi
uvicorn[standard]
